Updated the GUI to get more inputs
However I also noted that the first time the program runs, when there still isn't a json file containing the info, the program will get the inputs from the user,create the json file and then crash. The second time it seems the program runs normally. I still haven't identified the origin of this bug